Restructuring Subsystems
