--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Tue Jan 28 19:48:44 AWST 2020
-gradle.version=6.0.1
